requirements.txt
aiofiles==23.2.1
aiogram==3.13.0
aiohappyeyeballs==2.6.1
aiohttp==3.10.11
aiosignal==1.3.2
annotated-types==0.7.0
asgiref==3.8.1
async-timeout==5.0.1
asyncpg==0.30.0
attrs==25.3.0
certifi==2025.1.31
colorama==0.4.6
Django==5.2
djangorestframework==3.15.2
frozenlist==1.5.0
greenlet==3.1.1
idna==3.10
magic-filter==1.0.12
multidict==6.4.3
pillow==11.2.1
propcache==0.3.1
psycopg2-binary==2.9.10
pydantic==2.8.2
pydantic_core==2.20.1
python-dotenv==1.0.1
qrcode==8.1
SQLAlchemy==2.0.38
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.2
yarl==1.19.0
